refactor: Popup keyframes 모션으로 전환 및 닫힘 애니 제거
