+++ src/shared/components/v2/popup/Popup.tsx
@@ -1,8 +1,5 @@
 import { useEffect, type ReactNode } from 'react';
 
-import { useMountMotionPhase } from '@shared/hooks/useMountMotionPhase';
-
-import { POPUP_FADE_OUT_MS } from './constants';
 import * as styles from './Popup.css.ts';
 import ActionButton from '../button/actionButton/ActionButton';
 import IconButton from '../button/IconButton';
@@ -45,11 +42,6 @@ const Popup = ({
 }: PopupProps) => {
   const hasWeak = weakBtnText != null && weakBtnText !== '';
 
-  const { phase, containerRef, requestClose, handleTransitionEnd } =
-    useMountMotionPhase({
-      exitDurationMs: POPUP_FADE_OUT_MS,
-    });
-
   useEffect(() => {
     const previousOverflow = document.body.style.overflow;
     document.body.style.overflow = 'hidden';
@@ -66,7 +58,7 @@ const Popup = ({
           <button
             type="button"
             className={styles.textButton({ role: 'weak', layout: 'paired' })}
-            onClick={() => requestClose(onCancel)}
+            onClick={onCancel}
           >
             {weakBtnText}
           </button>
@@ -79,7 +71,7 @@ const Popup = ({
           })}
           disabled={confirmDisabled}
           aria-disabled={confirmDisabled || undefined}
-          onClick={() => requestClose(onConfirm)}
+          onClick={onConfirm}
         >
           {btnText}
         </button>
@@ -91,7 +83,7 @@ const Popup = ({
             name={sideIconName}
             size="M"
             className={styles.sideIconButton}
-            onClick={() => requestClose(onCancel)}
+            onClick={onCancel}
           />
         ) : null}
         <div className={styles.primaryButtonWrap}>
@@ -102,7 +94,7 @@ const Popup = ({
             leftIcon={btnIcon}
             fullWidth
             disabled={confirmDisabled}
-            onClick={() => requestClose(onConfirm)}
+            onClick={onConfirm}
           >
             {btnText}
           </ActionButton>
@@ -112,19 +104,13 @@ const Popup = ({
 
   return (
     <div className={styles.viewportLayer}>
+      <div className={styles.backdrop} onClick={onClose} aria-hidden="true" />
       <div
-        className={styles.backdrop({ phase })}
-        onClick={() => requestClose(onClose)}
-        aria-hidden="true"
-      />
-      <div
-        ref={containerRef}
-        className={styles.container({ phase })}
+        className={styles.container}
         role="dialog"
         aria-modal="true"
         aria-label={ariaLabel ?? '안내 팝업'}
         onClick={(e) => e.stopPropagation()}
-        onTransitionEnd={handleTransitionEnd}
       >
         {showCloseButton ? (
           <div className={styles.closeButton}>
@@ -132,7 +118,7 @@ const Popup = ({
               name="CloseFillGray"
               size="M"
               aria-label="닫기"
-              onClick={() => requestClose(onClose)}
+              onClick={onClose}
             />
           </div>
         ) : null}
